services/oil-spill-detector/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-cloud-firestore==2.16.0
google-auth==2.32.0
earthengine-api==0.1.405